+++ custom/plugins/zsh-autosuggestions/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## v0.7.1
+- Clear POSTDISPLAY instead of unsetting (#634)
+- Always reset async file descriptor after consuming it (#630)
+- Always use builtin `exec` (#628)
+- Add `history-beginning-search-*-end` widgets to clear widget list (#619)
+- Switch CI from Circle CI to GitHub Actions
+
+## v0.7.0
+- Enable asynchronous mode by default (#498)
+- No longer wrap user widgets starting with `autosuggest-` prefix (#496)
+- Fix a bug wrapping widgets that modify the buffer (#541)
+
+
 ## v0.6.4
 - Fix `vi-forward-char` triggering a bell when using it to accept a suggestion (#488)
 - New configuration option to skip completion suggestions when buffer matches a pattern (#487)